Move utility code to its own file, change FogProvider to a module, and include it
